refactor(core): move provider configs to provider modules
Move AnthropicConfig, GoogleAiConfig, and OpenAICompatibleConfig
from config.rs to their respective provider modules.
