+++ crates/rullm-core/src/config.rs
@@ -2,6 +2,8 @@ use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
 use std::time::Duration;
 
+use crate::providers::{AnthropicConfig, GoogleAiConfig, OpenAICompatibleConfig, OpenAIConfig};
+
 /// Configuration trait for LLM providers
 pub trait ProviderConfig: Send + Sync {
     /// Get the API key for this provider
@@ -81,249 +83,6 @@ impl ProviderConfig for HttpProviderConfig {
         if !self.base_url.starts_with("http://") && !self.base_url.starts_with("https://") {
             return Err(crate::error::LlmError::configuration(
                 "Base URL must be a valid HTTP/HTTPS URL",
-            ));
-        }
-
-        Ok(())
-    }
-}
-
-/// OpenAI-compatible configuration (supports OpenAI, Groq, OpenRouter, etc.)
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct OpenAICompatibleConfig {
-    pub api_key: String,
-    pub organization: Option<String>,
-    pub project: Option<String>,
-    pub base_url: Option<String>,
-    pub timeout_seconds: u64,
-}
-
-/// Type alias for backwards compatibility
-pub type OpenAIConfig = OpenAICompatibleConfig;
-
-impl OpenAICompatibleConfig {
-    pub fn new(api_key: impl Into<String>) -> Self {
-        Self {
-            api_key: api_key.into(),
-            organization: None,
-            project: None,
-            base_url: None,
-            timeout_seconds: 30,
-        }
-    }
-
-    pub fn groq(api_key: impl Into<String>) -> Self {
-        Self {
-            api_key: api_key.into(),
-            organization: None,
-            project: None,
-            base_url: Some("https://api.groq.com/openai/v1".to_string()),
-            timeout_seconds: 30,
-        }
-    }
-
-    pub fn openrouter(api_key: impl Into<String>) -> Self {
-        Self {
-            api_key: api_key.into(),
-            organization: None,
-            project: None,
-            base_url: Some("https://openrouter.ai/api/v1".to_string()),
-            timeout_seconds: 30,
-        }
-    }
-
-    pub fn with_organization(mut self, org: impl Into<String>) -> Self {
-        self.organization = Some(org.into());
-        self
-    }
-
-    pub fn with_project(mut self, project: impl Into<String>) -> Self {
-        self.project = Some(project.into());
-        self
-    }
-
-    pub fn with_base_url(mut self, base_url: impl Into<String>) -> Self {
-        self.base_url = Some(base_url.into());
-        self
-    }
-}
-
-impl ProviderConfig for OpenAICompatibleConfig {
-    fn api_key(&self) -> &str {
-        &self.api_key
-    }
-
-    fn base_url(&self) -> &str {
-        self.base_url
-            .as_deref()
-            .unwrap_or("https://api.openai.com/v1")
-    }
-
-    fn timeout(&self) -> Duration {
-        Duration::from_secs(self.timeout_seconds)
-    }
-
-    fn headers(&self) -> HashMap<String, String> {
-        let mut headers = HashMap::new();
-        headers.insert(
-            "Authorization".to_string(),
-            format!("Bearer {}", self.api_key),
-        );
-        headers.insert("Content-Type".to_string(), "application/json".to_string());
-
-        if let Some(org) = &self.organization {
-            headers.insert("OpenAI-Organization".to_string(), org.clone());
-        }
-
-        if let Some(project) = &self.project {
-            headers.insert("OpenAI-Project".to_string(), project.clone());
-        }
-
-        headers
-    }
-
-    fn validate(&self) -> Result<(), crate::error::LlmError> {
-        if self.api_key.is_empty() {
-            return Err(crate::error::LlmError::configuration("API key is required"));
-        }
-
-        // Relaxed validation: don't require 'sk-' prefix since Groq and OpenRouter use different formats
-        // OpenAI keys start with 'sk-', Groq uses 'gsk_', OpenRouter uses different format
-
-        Ok(())
-    }
-}
-
-/// Anthropic-specific configuration
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct AnthropicConfig {
-    pub api_key: String,
-    pub base_url: Option<String>,
-    pub timeout_seconds: u64,
-    /// Whether to use OAuth authentication (Bearer token) instead of API key (x-api-key)
-    #[serde(default)]
-    pub use_oauth: bool,
-}
-
-impl AnthropicConfig {
-    pub fn new(api_key: impl Into<String>) -> Self {
-        Self {
-            api_key: api_key.into(),
-            base_url: None,
-            timeout_seconds: 30,
-            use_oauth: false,
-        }
-    }
-
-    pub fn with_base_url(mut self, base_url: impl Into<String>) -> Self {
-        self.base_url = Some(base_url.into());
-        self
-    }
-
-    pub fn with_oauth(mut self, use_oauth: bool) -> Self {
-        self.use_oauth = use_oauth;
-        self
-    }
-}
-
-impl ProviderConfig for AnthropicConfig {
-    fn api_key(&self) -> &str {
-        &self.api_key
-    }
-
-    fn base_url(&self) -> &str {
-        self.base_url
-            .as_deref()
-            .unwrap_or("https://api.anthropic.com")
-    }
-
-    fn timeout(&self) -> Duration {
-        Duration::from_secs(self.timeout_seconds)
-    }
-
-    fn headers(&self) -> HashMap<String, String> {
-        let mut headers = HashMap::new();
-
-        if self.use_oauth {
-            // OAuth: use Bearer token + required beta headers
-            // Note: OpenCode doesn't send anthropic-version for OAuth requests
-            headers.insert(
-                "Authorization".to_string(),
-                format!("Bearer {}", self.api_key),
-            );
-            headers.insert(
-                "anthropic-beta".to_string(),
-                "oauth-2025-04-20,claude-code-20250219,interleaved-thinking-2025-05-14,fine-grained-tool-streaming-2025-05-14".to_string(),
-            );
-        } else {
-            // API key: use x-api-key header
-            headers.insert("x-api-key".to_string(), self.api_key.clone());
-            headers.insert("anthropic-version".to_string(), "2023-06-01".to_string());
-        }
-
-        headers.insert("Content-Type".to_string(), "application/json".to_string());
-        headers
-    }
-
-    fn validate(&self) -> Result<(), crate::error::LlmError> {
-        if self.api_key.is_empty() {
-            return Err(crate::error::LlmError::configuration(
-                "Anthropic API key is required",
-            ));
-        }
-
-        Ok(())
-    }
-}
-
-/// Google AI configuration
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct GoogleAiConfig {
-    pub api_key: String,
-    pub base_url: Option<String>,
-    pub timeout_seconds: u64,
-}
-
-impl GoogleAiConfig {
-    pub fn new(api_key: impl Into<String>) -> Self {
-        Self {
-            api_key: api_key.into(),
-            base_url: None,
-            timeout_seconds: 30,
-        }
-    }
-
-    pub fn with_base_url(mut self, base_url: impl Into<String>) -> Self {
-        self.base_url = Some(base_url.into());
-        self
-    }
-}
-
-impl ProviderConfig for GoogleAiConfig {
-    fn api_key(&self) -> &str {
-        &self.api_key
-    }
-
-    fn base_url(&self) -> &str {
-        self.base_url
-            .as_deref()
-            .unwrap_or("https://generativelanguage.googleapis.com/v1beta")
-    }
-
-    fn timeout(&self) -> Duration {
-        Duration::from_secs(self.timeout_seconds)
-    }
-
-    fn headers(&self) -> HashMap<String, String> {
-        let mut headers = HashMap::new();
-        headers.insert("Content-Type".to_string(), "application/json".to_string());
-        headers
-    }
-
-    fn validate(&self) -> Result<(), crate::error::LlmError> {
-        if self.api_key.is_empty() {
-            return Err(crate::error::LlmError::configuration(
-                "Google AI API key is required",
             ));
         }
