+++ crates/rullm-core/src/lib.rs
@@ -157,10 +157,7 @@ pub mod utils;
 // Concrete client exports
 pub use providers::{AnthropicClient, GoogleClient, OpenAIClient, OpenAICompatibleProvider};
 
-pub use config::{
-    AnthropicConfig, ConfigBuilder, GoogleAiConfig, OpenAICompatibleConfig, OpenAIConfig,
-    ProviderConfig,
-};
+pub use config::{ConfigBuilder, HttpProviderConfig, ProviderConfig};
 pub use error::LlmError;
 pub use utils::sse::sse_lines;
