+++ crates/rullm-core/src/providers/anthropic/client.rs
@@ -1,5 +1,6 @@
+use super::config::AnthropicConfig;
 use super::types::*;
-use crate::config::{AnthropicConfig, ProviderConfig};
+use crate::config::ProviderConfig;
 use crate::error::LlmError;
 use crate::utils::sse::sse_lines;
 use futures::Stream;
