+++ crates/rullm-core/src/providers/anthropic/mod.rs
@@ -24,7 +24,9 @@
 //! ```
 
 pub mod client;
+pub mod config;
 pub mod types;
 
 pub use client::AnthropicClient;
+pub use config::AnthropicConfig;
 pub use types::*;
